virtual_player/requirements.txt
python-socketio
flask
flask-cors
autopep8
requests
websocket-client
ipykernel
